Подразобрался с варнингами, теперь их меньше - в основном пофиксил, кое где задавил

+++ tokenizer/utils.h
@@ -410,9 +410,9 @@ template<typename CharType>
 inline
 int charToDigit(CharType ch)
 {
-    if (ch>=(CharType)'0' && ch<=(CharType)'9') return ch-(CharType)'0';
-    if (ch>=(CharType)'A' && ch<=(CharType)'F') return ch-(CharType)'A'+10;
-    if (ch>=(CharType)'a' && ch<=(CharType)'f') return ch-(CharType)'a'+10;
+    if (ch>=(CharType)'0' && ch<=(CharType)'9') return (int)(ch-(CharType)'0');
+    if (ch>=(CharType)'A' && ch<=(CharType)'F') return (int)(ch-(CharType)'A'+10);
+    if (ch>=(CharType)'a' && ch<=(CharType)'f') return (int)(ch-(CharType)'a'+10);
     return -1;
 }
